remove Box locking
As it was, it didn't do what it promised. The MOC would've probably had to be saved first if other Core Data clients access the store.

More importantly, this introduced a concept into the Domain just for the sake of transactional consistency. This is a possible Domain concept, not being able to add Items to a Box. It was introduced to solve an implementation problem only, though.